Extract CSP logic to increase maintainability score
