+++ src/Geopilot.Api/WebApplicationExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using Geopilot.Api.Pipeline;
 using Geopilot.Api.Pipeline.Process;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Cryptography;
 
 namespace Geopilot.Api;
 
@@ -51,5 +52,43 @@ public static class WebApplicationExtensions
         {
             throw new InvalidOperationException($"Invalid pipeline processes found:{Environment.NewLine}{string.Join(Environment.NewLine, invalidProcessesErrors)}");
         }
+    }
+
+    /// <summary>
+    /// Maps the SPA fallback route with CSP headers and nonce-based script/style injection.
+    /// </summary>
+    public static void MapSpaFallback(this WebApplication app, IConfiguration configuration)
+    {
+        ArgumentNullException.ThrowIfNull(app, nameof(app));
+
+        var indexHtmlPath = !string.IsNullOrEmpty(app.Environment.WebRootPath) ? Path.Combine(app.Environment.WebRootPath, "index.html") : null;
+        if (string.IsNullOrEmpty(indexHtmlPath) || !File.Exists(indexHtmlPath))
+        {
+            return;
+        }
+
+        var indexHtmlTemplate = File.ReadAllText(indexHtmlPath);
+        var authorityOrigin = new Uri(configuration["Auth:Authority"]!).GetLeftPart(UriPartial.Authority);
+        var blobEndpoint = configuration["CloudStorage:BlobEndpoint"];
+        if (!string.IsNullOrWhiteSpace(blobEndpoint))
+        {
+            if (!Uri.TryCreate(blobEndpoint, UriKind.Absolute, out var blobUri))
+                throw new InvalidOperationException($"CloudStorage:BlobEndpoint '{blobEndpoint}' is not a valid absolute URI.");
+            blobEndpoint = blobUri.GetLeftPart(UriPartial.Authority);
+        }
+
+        var connectSrc = string.IsNullOrWhiteSpace(blobEndpoint)
+            ? $"'self' {authorityOrigin}"
+            : $"'self' {authorityOrigin} {blobEndpoint}";
+
+        app.MapFallback(async context =>
+        {
+            var nonce = Convert.ToBase64String(RandomNumberGenerator.GetBytes(16));
+            context.Response.Headers.Append(
+                "Content-Security-Policy",
+                $"default-src 'self'; script-src 'strict-dynamic' 'nonce-{nonce}'; style-src 'nonce-{nonce}'; object-src 'none'; base-uri 'none'; connect-src {connectSrc}; form-action 'self'; frame-ancestors 'none'; require-trusted-types-for 'script';");
+            context.Response.ContentType = "text/html";
+            await context.Response.WriteAsync(indexHtmlTemplate.Replace("__CSP_NONCE__", nonce));
+        }).AllowAnonymous();
     }
 }
